fix(menu): fix menu error caused by composition of server/client components
menu-items.tsx was a async server component, this worked fine on the desktop rendered version (from
another server component, but not on the mobile version. Since the mobile version is rendered inside
a drawer (client component), rendering an async server component inside of it caused an error when
opening the menu on mobile.